+++ src/maconn.ts
@@ -1,9 +1,9 @@
-import type { MultiaddrConnection, MultiaddrConnectionTimeline } from '@libp2p/interface-connection'
 import { logger } from '@libp2p/logger'
+import { nopSink, nopSource } from './util.js'
+import type { MultiaddrConnection, MultiaddrConnectionTimeline } from '@libp2p/interface-connection'
+import type { CounterGroup } from '@libp2p/interface-metrics'
 import type { Multiaddr } from '@multiformats/multiaddr'
 import type { Source, Sink } from 'it-stream-types'
-
-import { nopSink, nopSource } from './util.js'
 
 const log = logger('libp2p:webrtc:connection')
 
@@ -22,6 +22,11 @@ interface WebRTCMultiaddrConnectionInit {
    * Holds the relevant events timestamps of the connection
    */
   timeline: MultiaddrConnectionTimeline
+
+  /**
+   * Optional metrics counter group for this connection
+   */
+  metrics?: CounterGroup
 }
 
 export class WebRTCMultiaddrConnection implements MultiaddrConnection {
@@ -41,29 +46,40 @@ export class WebRTCMultiaddrConnection implements MultiaddrConnection {
   timeline: MultiaddrConnectionTimeline
 
   /**
+   * Optional metrics counter group for this connection
+   */
+  metrics?: CounterGroup
+
+  /**
    * The stream source, a no-op as the transport natively supports multiplexing
    */
-  source: Source<Uint8Array> = nopSource
+  source: AsyncGenerator<Uint8Array, any, unknown> = nopSource()
 
   /**
    * The stream destination, a no-op as the transport natively supports multiplexing
    */
-  sink: Sink<Uint8Array, Promise<void>> = nopSink
+  sink: Sink<Source<Uint8Array>, Promise<void>> = nopSink
 
   constructor (init: WebRTCMultiaddrConnectionInit) {
     this.remoteAddr = init.remoteAddr
     this.timeline = init.timeline
     this.peerConnection = init.peerConnection
+
+    this.peerConnection.onconnectionstatechange = () => {
+      if (this.peerConnection.connectionState === 'closed' || this.peerConnection.connectionState === 'disconnected' || this.peerConnection.connectionState === 'failed') {
+        this.timeline.close = Date.now()
+      }
+    }
   }
 
   async close (err?: Error | undefined): Promise<void> {
     if (err !== undefined) {
       log.error('error closing connection', err)
     }
-
     log.trace('closing connection')
 
     this.timeline.close = Date.now()
     this.peerConnection.close()
+    this.metrics?.increment({ close: true })
   }
 }