+++ src/private-to-private/handler.ts
@@ -1,14 +1,13 @@
-import type { IncomingStreamData } from '@libp2p/interface-registrar'
+import { logger } from '@libp2p/logger'
+import { abortableDuplex } from 'abortable-iterator'
 import { pbStream } from 'it-pb-stream'
 import pDefer, { type DeferredPromise } from 'p-defer'
-import { TimeoutController } from 'timeout-abort-controller'
-import { readCandidatesUntilConnected, resolveOnConnected } from './util.js'
-import * as pb from './pb/index.js'
-import { abortableDuplex } from 'abortable-iterator'
-import { logger } from '@libp2p/logger'
-import type { Stream } from '@libp2p/interface-connection'
-import type { StreamMuxerFactory } from '@libp2p/interface-stream-muxer'
 import { DataChannelMuxerFactory } from '../muxer.js'
+import { Message } from './pb/message.js'
+import { readCandidatesUntilConnected, resolveOnConnected } from './util.js'
+import type { Stream } from '@libp2p/interface-connection'
+import type { IncomingStreamData } from '@libp2p/interface-registrar'
+import type { StreamMuxerFactory } from '@libp2p/interface-stream-muxer'
 
 const DEFAULT_TIMEOUT = 30 * 1000
 
@@ -16,10 +15,9 @@ const log = logger('libp2p:webrtc:peer')
 
 export type IncomingStreamOpts = { rtcConfiguration?: RTCConfiguration } & IncomingStreamData
 
-export async function handleIncomingStream ({ rtcConfiguration, stream: rawStream }: IncomingStreamOpts): Promise<[RTCPeerConnection, StreamMuxerFactory]> {
-  const timeoutController = new TimeoutController(DEFAULT_TIMEOUT)
-  const signal = timeoutController.signal
-  const stream = pbStream(abortableDuplex(rawStream, timeoutController.signal)).pb(pb.Message)
+export async function handleIncomingStream ({ rtcConfiguration, stream: rawStream }: IncomingStreamOpts): Promise<{ pc: RTCPeerConnection, muxerFactory: StreamMuxerFactory, remoteAddress: string }> {
+  const signal = AbortSignal.timeout(DEFAULT_TIMEOUT)
+  const stream = pbStream(abortableDuplex(rawStream, signal)).pb(Message)
   const pc = new RTCPeerConnection(rtcConfiguration)
   const muxerFactory = new DataChannelMuxerFactory(pc)
 
@@ -32,7 +30,7 @@ export async function handleIncomingStream ({ rtcConfiguration, stream: rawStrea
     answerSentPromise.promise.then(
       () => {
         stream.write({
-          type: pb.Message.Type.ICE_CANDIDATE,
+          type: Message.Type.ICE_CANDIDATE,
           data: (candidate != null) ? JSON.stringify(candidate.toJSON()) : ''
         })
       },
@@ -46,7 +44,7 @@ export async function handleIncomingStream ({ rtcConfiguration, stream: rawStrea
 
   // read an SDP offer
   const pbOffer = await stream.read()
-  if (pbOffer.type !== pb.Message.Type.SDP_OFFER) {
+  if (pbOffer.type !== Message.Type.SDP_OFFER) {
     throw new Error(`expected message type SDP_OFFER, received: ${pbOffer.type ?? 'undefined'} `)
   }
   const offer = new RTCSessionDescription({
@@ -66,7 +64,7 @@ export async function handleIncomingStream ({ rtcConfiguration, stream: rawStrea
     throw new Error('Failed to create answer')
   })
   // write the answer to the remote
-  stream.write({ type: pb.Message.Type.SDP_ANSWER, data: answer.sdp })
+  stream.write({ type: Message.Type.SDP_ANSWER, data: answer.sdp })
 
   await pc.setLocalDescription(answer).catch(err => {
     log.error('could not execute setLocalDescription', err)
@@ -78,7 +76,10 @@ export async function handleIncomingStream ({ rtcConfiguration, stream: rawStrea
 
   // wait until candidates are connected
   await readCandidatesUntilConnected(connectedPromise, pc, stream)
-  return [pc, muxerFactory]
+
+  const remoteAddress = parseRemoteAddress(pc.currentRemoteDescription?.sdp ?? '')
+
+  return { pc, muxerFactory, remoteAddress }
 }
 
 export interface ConnectOptions {
@@ -87,9 +88,8 @@ export interface ConnectOptions {
   rtcConfiguration?: RTCConfiguration
 }
 
-export async function initiateConnection ({ rtcConfiguration, signal, stream: rawStream }: ConnectOptions): Promise<[RTCPeerConnection, StreamMuxerFactory]> {
-  const stream = pbStream(abortableDuplex(rawStream, signal)).pb(pb.Message)
-
+export async function initiateConnection ({ rtcConfiguration, signal, stream: rawStream }: ConnectOptions): Promise<{ pc: RTCPeerConnection, muxerFactory: StreamMuxerFactory, remoteAddress: string }> {
+  const stream = pbStream(abortableDuplex(rawStream, signal)).pb(Message)
   // setup peer connection
   const pc = new RTCPeerConnection(rtcConfiguration)
   const muxerFactory = new DataChannelMuxerFactory(pc)
@@ -107,14 +107,14 @@ export async function initiateConnection ({ rtcConfiguration, signal, stream: ra
   // peer
   pc.onicecandidate = ({ candidate }) => {
     stream.write({
-      type: pb.Message.Type.ICE_CANDIDATE,
+      type: Message.Type.ICE_CANDIDATE,
       data: (candidate != null) ? JSON.stringify(candidate.toJSON()) : ''
     })
   }
   // create an offer
   const offerSdp = await pc.createOffer()
   // write the offer to the stream
-  stream.write({ type: pb.Message.Type.SDP_OFFER, data: offerSdp.sdp })
+  stream.write({ type: Message.Type.SDP_OFFER, data: offerSdp.sdp })
   // set offer as local description
   await pc.setLocalDescription(offerSdp).catch(err => {
     log.error('could not execute setLocalDescription', err)
@@ -123,7 +123,7 @@ export async function initiateConnection ({ rtcConfiguration, signal, stream: ra
 
   // read answer
   const answerMessage = await stream.read()
-  if (answerMessage.type !== pb.Message.Type.SDP_ANSWER) {
+  if (answerMessage.type !== Message.Type.SDP_ANSWER) {
     throw new Error('remote should send an SDP answer')
   }
 
@@ -135,5 +135,21 @@ export async function initiateConnection ({ rtcConfiguration, signal, stream: ra
 
   await readCandidatesUntilConnected(connectedPromise, pc, stream)
   channel.close()
-  return [pc, muxerFactory]
+
+  const remoteAddress = parseRemoteAddress(pc.currentRemoteDescription?.sdp ?? '')
+
+  return { pc, muxerFactory, remoteAddress }
+}
+
+function parseRemoteAddress (sdp: string): string {
+  // 'a=candidate:1746876089 1 udp 2113937151 0614fbad-b...ocal 54882 typ host generation 0 network-cost 999'
+  const candidateLine = sdp.split('\r\n').filter(line => line.startsWith('a=candidate')).pop()
+  const candidateParts = candidateLine?.split(' ')
+
+  if (candidateLine == null || candidateParts == null || candidateParts.length < 5) {
+    log('could not parse remote address from', candidateLine)
+    return '/webrtc'
+  }
+
+  return `/dnsaddr/${candidateParts[4]}/${candidateParts[2]}/${candidateParts[3]}/webrtc`
 }