+++ test/transport.browser.spec.ts
@@ -1,13 +1,14 @@
 /* eslint-disable @typescript-eslint/no-floating-promises */
 
-import * as underTest from './../src/transport.js'
-import { expectError } from './util.js'
-import { UnimplementedError } from './../src/error.js'
-import { mockUpgrader } from '@libp2p/interface-mocks'
-import { CreateListenerOptions, symbol } from '@libp2p/interface-transport'
-import { multiaddr, Multiaddr } from '@multiformats/multiaddr'
+import { mockMetrics, mockUpgrader } from '@libp2p/interface-mocks'
+import { type CreateListenerOptions, symbol } from '@libp2p/interface-transport'
 import { createEd25519PeerId } from '@libp2p/peer-id-factory'
+import { multiaddr, type Multiaddr } from '@multiformats/multiaddr'
 import { expect, assert } from 'aegir/chai'
+import { UnimplementedError } from './../src/error.js'
+import * as underTest from './../src/private-to-public/transport.js'
+import { expectError } from './util.js'
+import type { Metrics } from '@libp2p/interface-metrics'
 
 function ignoredDialOption (): CreateListenerOptions {
   const upgrader = mockUpgrader({})
@@ -15,11 +16,14 @@ function ignoredDialOption (): CreateListenerOptions {
 }
 
 describe('WebRTC Transport', () => {
+  let metrics: Metrics
   let components: underTest.WebRTCDirectTransportComponents
 
   before(async () => {
+    metrics = mockMetrics()()
     components = {
-      peerId: await createEd25519PeerId()
+      peerId: await createEd25519PeerId(),
+      metrics
     }
   })